Close socket on proxy shutdown
When TCP Proxy is shut down, underlying socket remains open.
Even though `tcpProxy.shutdown()` is called, it is still possible to
connect to the socket listening on a specified port.

To close NIO selector properly, one has to iterate over its keys and close a
channel assigned to each key.

Closes issue #12